requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.14
flake8==7.3.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
numpy==2.2.6
oauth==1.0.1
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2==2.9.10
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.35.0
win32_setctime==1.2.0
